+++ backend/src/DataFixtures/SystemVariablesFixtureLoader.php
@@ -1,4 +1,6 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace Neucore\DataFixtures;
 
@@ -8,6 +10,7 @@ use Neucore\Factory\RepositoryFactory;
 use Doctrine\Common\DataFixtures\FixtureInterface;
 use Doctrine\Common\Persistence\ObjectManager;
 
+/** @noinspection PhpUnused */
 class SystemVariablesFixtureLoader implements FixtureInterface
 {
     public function load(ObjectManager $manager)
@@ -17,43 +20,49 @@ class SystemVariablesFixtureLoader implements FixtureInterface
         $pathToImages = Application::ROOT_DIR . '/var';
         $imagePrefix = 'data:image/png;base64,';
         $vars = [
-            SystemVariable::ACCOUNT_DEACTIVATION_DELAY      => ['',  SystemVariable::SCOPE_SETTINGS],
-            SystemVariable::ALLOW_CHARACTER_DELETION        => ['0', SystemVariable::SCOPE_PUBLIC],
-            SystemVariable::ALLOW_LOGIN_MANAGED             => ['0', SystemVariable::SCOPE_SETTINGS],
-            SystemVariable::ESI_ERROR_LIMIT                 => ['',  SystemVariable::SCOPE_BACKEND],
-            SystemVariable::GROUPS_REQUIRE_VALID_TOKEN      => ['0', SystemVariable::SCOPE_PUBLIC],
-            SystemVariable::MAIL_INVALID_TOKEN_ACTIVE       => ['',  SystemVariable::SCOPE_SETTINGS],
-            SystemVariable::MAIL_INVALID_TOKEN_ALLIANCES    => ['',  SystemVariable::SCOPE_SETTINGS],
-            SystemVariable::MAIL_INVALID_TOKEN_BODY         => ['',  SystemVariable::SCOPE_SETTINGS],
-            SystemVariable::MAIL_INVALID_TOKEN_SUBJECT      => ['',  SystemVariable::SCOPE_SETTINGS],
-            SystemVariable::MAIL_CHARACTER                  => ['',  SystemVariable::SCOPE_SETTINGS],
-            SystemVariable::MAIL_TOKEN                      => ['',  SystemVariable::SCOPE_BACKEND],
-            SystemVariable::CUSTOMIZATION_DOCUMENT_TITLE    => ['Alliance Core Services', SystemVariable::SCOPE_PUBLIC],
-            SystemVariable::CUSTOMIZATION_DEFAULT_THEME     => ['Darkly', SystemVariable::SCOPE_PUBLIC],
-            SystemVariable::CUSTOMIZATION_WEBSITE           => [
+            SystemVariable::GROUPS_REQUIRE_VALID_TOKEN        => ['0', SystemVariable::SCOPE_SETTINGS],
+            SystemVariable::ACCOUNT_DEACTIVATION_DELAY        => ['',  SystemVariable::SCOPE_SETTINGS],
+            SystemVariable::ACCOUNT_DEACTIVATION_ALLIANCES    => ['',  SystemVariable::SCOPE_SETTINGS],
+            SystemVariable::ACCOUNT_DEACTIVATION_CORPORATIONS => ['',  SystemVariable::SCOPE_SETTINGS],
+            SystemVariable::ALLOW_CHARACTER_DELETION          => ['0', SystemVariable::SCOPE_PUBLIC],
+            SystemVariable::ALLOW_LOGIN_MANAGED               => ['0', SystemVariable::SCOPE_SETTINGS],
+            SystemVariable::ESI_ERROR_LIMIT                   => ['',  SystemVariable::SCOPE_BACKEND],
+            SystemVariable::MAIL_INVALID_TOKEN_ACTIVE         => ['',  SystemVariable::SCOPE_SETTINGS],
+            SystemVariable::MAIL_INVALID_TOKEN_ALLIANCES      => ['',  SystemVariable::SCOPE_SETTINGS],
+            SystemVariable::MAIL_INVALID_TOKEN_CORPORATIONS   => ['',  SystemVariable::SCOPE_SETTINGS],
+            SystemVariable::MAIL_INVALID_TOKEN_BODY           => ['',  SystemVariable::SCOPE_SETTINGS],
+            SystemVariable::MAIL_INVALID_TOKEN_SUBJECT        => ['',  SystemVariable::SCOPE_SETTINGS],
+            SystemVariable::MAIL_CHARACTER                    => ['',  SystemVariable::SCOPE_SETTINGS],
+            SystemVariable::MAIL_TOKEN                        => ['',  SystemVariable::SCOPE_BACKEND],
+            SystemVariable::CUSTOMIZATION_DOCUMENT_TITLE      => [
+                'Alliance Core Services',
+                SystemVariable::SCOPE_PUBLIC
+            ],
+            SystemVariable::CUSTOMIZATION_DEFAULT_THEME       => ['Darkly', SystemVariable::SCOPE_PUBLIC],
+            SystemVariable::CUSTOMIZATION_WEBSITE             => [
                 'https://github.com/tkhamez/neucore',
                 SystemVariable::SCOPE_PUBLIC
             ],
-            SystemVariable::CUSTOMIZATION_NAV_TITLE         => ['Neucore', SystemVariable::SCOPE_PUBLIC],
-            SystemVariable::CUSTOMIZATION_NAV_LOGO          => [
+            SystemVariable::CUSTOMIZATION_NAV_TITLE           => ['Neucore', SystemVariable::SCOPE_PUBLIC],
+            SystemVariable::CUSTOMIZATION_NAV_LOGO            => [
                 $imagePrefix . base64_encode((string) file_get_contents($pathToImages . '/logo_29.png')),
                 SystemVariable::SCOPE_PUBLIC
             ],
-            SystemVariable::CUSTOMIZATION_HOME_HEADLINE     => ['Core Services', SystemVariable::SCOPE_PUBLIC],
-            SystemVariable::CUSTOMIZATION_HOME_DESCRIPTION  => [
+            SystemVariable::CUSTOMIZATION_HOME_HEADLINE       => ['Core Services', SystemVariable::SCOPE_PUBLIC],
+            SystemVariable::CUSTOMIZATION_HOME_DESCRIPTION    => [
                 'An application to manage access for EVE Online players to external services.',
                 SystemVariable::SCOPE_PUBLIC
             ],
-            SystemVariable::CUSTOMIZATION_HOME_LOGO         => [
+            SystemVariable::CUSTOMIZATION_HOME_LOGO           => [
                 $imagePrefix . base64_encode((string) file_get_contents($pathToImages . '/logo_300.png')),
                 SystemVariable::SCOPE_PUBLIC
             ],
-            SystemVariable::CUSTOMIZATION_HOME_MARKDOWN     => ['', SystemVariable::SCOPE_PUBLIC],
-            SystemVariable::CUSTOMIZATION_FOOTER_TEXT       => [
+            SystemVariable::CUSTOMIZATION_HOME_MARKDOWN       => ['', SystemVariable::SCOPE_PUBLIC],
+            SystemVariable::CUSTOMIZATION_FOOTER_TEXT         => [
                 'Documentation is available on GitHub.',
                 SystemVariable::SCOPE_PUBLIC
             ],
-            SystemVariable::CUSTOMIZATION_GITHUB            => [
+            SystemVariable::CUSTOMIZATION_GITHUB              => [
                 'https://github.com/tkhamez/neucore',
                 SystemVariable::SCOPE_PUBLIC
             ],