+++ backend/tests/Functional/Command/DoctrineFixturesLoadTest.php
@@ -38,10 +38,10 @@ class DoctrineFixturesLoadTest extends ConsoleTestCase
         $vars = $repoFactory->getSystemVariableRepository()->findBy([], ['name' => 'asc']);
 
         $this->assertSame(16, count($roles)); // 15 from seed + 1 from setup
-        $this->assertSame(23, count($vars)); // 22 from seed + 1 from setup
+        $this->assertSame(26, count($vars)); // 25 from seed + 1 from setup
 
         // check that value was not changed
-        $this->assertSame(SystemVariable::ALLOW_CHARACTER_DELETION, $vars[1]->getName());
-        $this->assertSame('1', $vars[1]->getValue());
+        $this->assertSame(SystemVariable::ALLOW_CHARACTER_DELETION, $vars[3]->getName());
+        $this->assertSame('1', $vars[3]->getValue());
     }
 }